requirements.txt
faster-whisper>=1.2.1,<2.0.0
ctranslate2>=4.6.3,<5.0.0
requests>=2.32.5,<3.0.0
tqdm>=4.67.1,<5.0.0
pyyaml>=6.0.3,<7.0.0
psycopg2-binary>=2.9.0,<3.0.0
paramiko>=3.4.0,<4.0.0
fastapi>=0.111.0
pydantic>=2.0.0,<3.0.0
uvicorn[standard]>=0.29.0
tiktoken>=0.12.0
